# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ALEF DELTA ERP API",
        "description": "Production-ready ERP API for manpower supply management with GPS-verified attendance, Telegram integration, payroll automation, and asset tracking.",
        "contact": {
            "email": "admin@alefdelta.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.alefdelta.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/attendance/clock-in": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Clock in for a shift",
                "operationId": "b9e586060b624a3880a68dc9a2f66bb4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "schedule_id",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "schedule_id": {
                                        "type": "integer"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "initData": {
                                        "description": "Telegram initData",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Clocked in successfully"
                    },
                    "400": {
                        "description": "Clock in failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/attendance/clock-out": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Clock out from a shift",
                "operationId": "023cdbfeb9affb051e9b8076c941cbd3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "schedule_id",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "schedule_id": {
                                        "type": "integer"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Clocked out successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/attendance/logs": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Get attendance logs (Admin/Manager)",
                "operationId": "9493f91e4f39fe6d875dd73259332e60",
                "parameters": [
                    {
                        "name": "employee_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "site_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_verified",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of attendance logs"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/attendance/my-logs": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Get my attendance logs",
                "operationId": "81a4dbd3bdde4f421b06bd139d27560f",
                "responses": {
                    "200": {
                        "description": "List of my attendance logs"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/attendance/export": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Export attendance logs to Excel",
                "operationId": "d9cdba4b039ef6d50efd0634b532a43e",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "site_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excel file download"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token obtained from login endpoint",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}